requirements.txt
Cython
connexion >= 2.7.0
python_dateutil >= 2.7.3
setuptools >= 50.3.0
Werkzeug >= 1.0.1
pandas == 0.25.3
Flask-SQLAlchemy == 2.4.4